+++ demos/demo_working_copy.sh
@@ -7,9 +7,9 @@ new_tmp_dir
     jj git clone https://github.com/octocat/Hello-World
     cd Hello-World
     jj abandon test
-    jj branch forget test
+    jj branch forget --global test
     jj abandon octocat-patch-1
-    jj branch forget octocat-patch-1
+    jj branch forget --global octocat-patch-1
 }> /dev/null
 
 comment "We are in the octocat/Hello-World repo.